requirements.txt
python-dotenv
black
Flask
flask_socketio
azure-cognitiveservices-speech
eventlet
